Moved logic to controller
