+++ client/src/controllers/QAControllers/QAPaginatedListController.jsx
@@ -1,6 +1,7 @@
 import { useQuestion } from 'contexts';
 import { PaginatedList } from 'controllers';
-import { Answer, AnswerComment, Comment } from 'components/QAComponents';
+import { AnswerComment, Comment } from 'components/QAComponents';
+import { Answer } from 'controllers/QAControllers';
 import { getAnswerComments, getAnswers, getQuestionComments } from 'services/questionsServices';
 
 export function AnswerCommentsList({ answer_id, comments: count }) {
